# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Logo Rush game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GameStats {
  playersOnline: number;
  publicRooms: number;
  gamesInProgress: number;
}

export type PublicRoomStatus = typeof PublicRoomStatus[keyof typeof PublicRoomStatus];


export const PublicRoomStatus = {
  waiting: 'waiting',
  playing: 'playing',
} as const;

export interface PublicRoom {
  code: string;
  name: string;
  themeId: string;
  hostName: string;
  playerCount: number;
  maxPlayers: number;
  roundCount: number;
  roundDuration: number;
  status: PublicRoomStatus;
}

export interface Theme {
  id: string;
  nameFr: string;
  nameEn: string;
  aspectW: number;
  aspectH: number;
}

export type LogoDifficulty = typeof LogoDifficulty[keyof typeof LogoDifficulty];


export const LogoDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Logo {
  id: string;
  themeId: string;
  answerFr: string;
  answerEn: string;
  aliasesFr: string[];
  aliasesEn: string[];
  category?: string;
  difficulty: LogoDifficulty;
  imageUrl: string;
}

export interface SoloLeaderboardEntry {
  id: number;
  nickname: string;
  score: number;
  themeId: string;
  roundCount: number;
  roundDuration: number;
  createdAt: string;
}

export type SoloScoreInputRoundCount = typeof SoloScoreInputRoundCount[keyof typeof SoloScoreInputRoundCount];


export const SoloScoreInputRoundCount = {
  NUMBER_5: 5,
  NUMBER_10: 10,
  NUMBER_15: 15,
  NUMBER_20: 20,
} as const;

export type SoloScoreInputRoundDuration = typeof SoloScoreInputRoundDuration[keyof typeof SoloScoreInputRoundDuration];


export const SoloScoreInputRoundDuration = {
  NUMBER_15: 15,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export interface SoloScoreInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  nickname: string;
  /**
     * @minimum 0
     * @maximum 20000
     */
  score: number;
  themeId: string;
  roundCount: SoloScoreInputRoundCount;
  roundDuration: SoloScoreInputRoundDuration;
}

export type SoloRoundDifficulty = typeof SoloRoundDifficulty[keyof typeof SoloRoundDifficulty];


export const SoloRoundDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface SoloRound {
  token: string;
  themeId: string;
  category?: string;
  difficulty: SoloRoundDifficulty;
  imageUrl: string;
}

export type SoloGuessInputLocale = typeof SoloGuessInputLocale[keyof typeof SoloGuessInputLocale];


export const SoloGuessInputLocale = {
  fr: 'fr',
  en: 'en',
} as const;

export interface SoloGuessInput {
  /** @minLength 1 */
  token: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  guess: string;
  locale?: SoloGuessInputLocale;
}

export interface SoloGuessResult {
  correct: boolean;
  answer?: string;
}

export type SoloRevealInputLocale = typeof SoloRevealInputLocale[keyof typeof SoloRevealInputLocale];


export const SoloRevealInputLocale = {
  fr: 'fr',
  en: 'en',
} as const;

export interface SoloRevealInput {
  /** @minLength 1 */
  token: string;
  locale?: SoloRevealInputLocale;
}

export interface SoloRevealResult {
  answer: string;
}

export type LogoReportInputReason = typeof LogoReportInputReason[keyof typeof LogoReportInputReason];


export const LogoReportInputReason = {
  unknown: 'unknown',
  incorrect: 'incorrect',
  other: 'other',
} as const;

export interface LogoReportInput {
  /**
     * @minLength 1
     * @maxLength 64
     */
  logoId: string;
  reason: LogoReportInputReason;
}

export interface LogoReportResult {
  ok: boolean;
}

export type GetSoloLeaderboardParams = {
themeId: string;
roundCount: GetSoloLeaderboardRoundCount;
roundDuration: GetSoloLeaderboardRoundDuration;
};

export type GetSoloLeaderboardRoundCount = typeof GetSoloLeaderboardRoundCount[keyof typeof GetSoloLeaderboardRoundCount];


export const GetSoloLeaderboardRoundCount = {
  NUMBER_5: 5,
  NUMBER_10: 10,
  NUMBER_15: 15,
  NUMBER_20: 20,
} as const;

export type GetSoloLeaderboardRoundDuration = typeof GetSoloLeaderboardRoundDuration[keyof typeof GetSoloLeaderboardRoundDuration];


export const GetSoloLeaderboardRoundDuration = {
  NUMBER_15: 15,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export type ListSoloRoundsParams = {
themeId: string;
};