# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Logo Rush game API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: game
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /game/stats:
    get:
      operationId: getGameStats
      tags: [game]
      summary: Live platform activity
      responses:
        "200":
          description: Current activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStats"
  /game/rooms:
    get:
      operationId: listPublicRooms
      tags: [game]
      summary: Public rooms currently accepting players
      responses:
        "200":
          description: Public rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicRoom"
  /game/themes:
    get:
      operationId: listThemes
      tags: [game]
      summary: Selectable game themes (brands, football clubs, movie posters, video games...)
      responses:
        "200":
          description: Enabled themes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theme"
  /game/logos:
    get:
      operationId: listSoloLogos
      tags: [game]
      summary: Full catalog dump across all themes, for the internal /logo-audit QA tool only
      responses:
        "200":
          description: Logos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Logo"
  /game/solo-leaderboard:
    get:
      operationId: getSoloLeaderboard
      tags: [game]
      summary: Top five solo scores for one game mode
      parameters:
        - name: themeId
          in: query
          required: true
          schema:
            type: string
        - name: roundCount
          in: query
          required: true
          schema:
            type: integer
            enum: [5, 10, 15, 20]
        - name: roundDuration
          in: query
          required: true
          schema:
            type: integer
            enum: [15, 20, 30]
      responses:
        "200":
          description: Best scores for this mode
          content:
            application/json:
              schema:
                type: array
                maxItems: 5
                items:
                  $ref: "#/components/schemas/SoloLeaderboardEntry"
    post:
      operationId: submitSoloScore
      tags: [game]
      summary: Save a completed solo game score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoloScoreInput"
      responses:
        "201":
          description: Updated leaderboard for this mode
          content:
            application/json:
              schema:
                type: array
                maxItems: 5
                items:
                  $ref: "#/components/schemas/SoloLeaderboardEntry"
  /game/solo/logos:
    get:
      operationId: listSoloRounds
      tags: [game]
      summary: Playable solo round pool (answers withheld until guessed or revealed)
      parameters:
        - name: themeId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Round pool
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SoloRound"
  /game/solo/guess:
    post:
      operationId: submitSoloGuess
      tags: [game]
      summary: Check a solo round guess server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoloGuessInput"
      responses:
        "200":
          description: Guess result; answer is only included when correct
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoloGuessResult"
  /game/solo/reveal:
    post:
      operationId: revealSoloRound
      tags: [game]
      summary: Reveal the answer for a solo round whose timer ran out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoloRevealInput"
      responses:
        "200":
          description: Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoloRevealResult"
  /game/logo-image/{token}:
    get:
      operationId: getLogoImage
      tags: [game]
      summary: Proxy a round's logo image without exposing its source domain
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Logo image
          content:
            image/*:
              schema:
                type: string
                format: binary
  /game/logo-reports:
    post:
      operationId: reportLogo
      tags: [game]
      summary: Flag a logo as unrecognizable or incorrect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoReportInput"
      responses:
        "201":
          description: Report recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoReportResult"
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    GameStats:
      type: object
      required: [playersOnline, publicRooms, gamesInProgress]
      properties:
        playersOnline:
          type: integer
        publicRooms:
          type: integer
        gamesInProgress:
          type: integer
    PublicRoom:
      type: object
      required: [code, name, themeId, hostName, playerCount, maxPlayers, roundCount, roundDuration, status]
      properties:
        code:
          type: string
        name:
          type: string
        themeId:
          type: string
        hostName:
          type: string
        playerCount:
          type: integer
        maxPlayers:
          type: integer
        roundCount:
          type: integer
        roundDuration:
          type: integer
        status:
          type: string
          enum: [waiting, playing]
    Theme:
      type: object
      required: [id, nameFr, nameEn, aspectW, aspectH]
      properties:
        id:
          type: string
        nameFr:
          type: string
        nameEn:
          type: string
        aspectW:
          type: integer
        aspectH:
          type: integer
    Logo:
      type: object
      required: [id, themeId, answerFr, answerEn, aliasesFr, aliasesEn, difficulty, imageUrl]
      properties:
        id:
          type: string
        themeId:
          type: string
        answerFr:
          type: string
        answerEn:
          type: string
        aliasesFr:
          type: array
          items:
            type: string
        aliasesEn:
          type: array
          items:
            type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        imageUrl:
          type: string
    SoloLeaderboardEntry:
      type: object
      required: [id, nickname, score, themeId, roundCount, roundDuration, createdAt]
      properties:
        id:
          type: integer
        nickname:
          type: string
        score:
          type: integer
        themeId:
          type: string
        roundCount:
          type: integer
        roundDuration:
          type: integer
        createdAt:
          type: string
          format: date-time
    SoloScoreInput:
      type: object
      required: [nickname, score, themeId, roundCount, roundDuration]
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 20
        score:
          type: integer
          minimum: 0
          maximum: 20000
        themeId:
          type: string
        roundCount:
          type: integer
          enum: [5, 10, 15, 20]
        roundDuration:
          type: integer
          enum: [15, 20, 30]
    SoloRound:
      type: object
      required: [token, themeId, difficulty, imageUrl]
      properties:
        token:
          type: string
        themeId:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        imageUrl:
          type: string
    SoloGuessInput:
      type: object
      required: [token, guess]
      properties:
        token:
          type: string
          minLength: 1
        guess:
          type: string
          minLength: 1
          maxLength: 80
        locale:
          type: string
          enum: [fr, en]
    SoloGuessResult:
      type: object
      required: [correct]
      properties:
        correct:
          type: boolean
        answer:
          type: string
    SoloRevealInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
        locale:
          type: string
          enum: [fr, en]
    SoloRevealResult:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
    LogoReportInput:
      type: object
      required: [logoId, reason]
      properties:
        logoId:
          type: string
          minLength: 1
          maxLength: 64
        reason:
          type: string
          enum: [unknown, incorrect, other]
    LogoReportResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean