# lib/api-zod/src/generated/types/publicRoom.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Logo Rush game API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicRoomStatus } from './publicRoomStatus';

export interface PublicRoom {
  code: string;
  name: string;
  themeId: string;
  hostName: string;
  playerCount: number;
  maxPlayers: number;
  roundCount: number;
  roundDuration: number;
  status: PublicRoomStatus;
}