# lib/api-zod/src/generated/types/soloScoreInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Logo Rush game API
 * OpenAPI spec version: 0.1.0
 */
import type { SoloScoreInputRoundCount } from './soloScoreInputRoundCount';
import type { SoloScoreInputRoundDuration } from './soloScoreInputRoundDuration';

export interface SoloScoreInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  nickname: string;
  /**
     * @minimum 0
     * @maximum 20000
     */
  score: number;
  themeId: string;
  roundCount: SoloScoreInputRoundCount;
  roundDuration: SoloScoreInputRoundDuration;
}