# lib/api-zod/src/generated/types/theme.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Logo Rush game API
 * OpenAPI spec version: 0.1.0
 */

export interface Theme {
  id: string;
  nameFr: string;
  nameEn: string;
  aspectW: number;
  aspectH: number;
}